requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.3.0
celery>=5.5.0
django-celery-beat>=2.8.0
django-celery-results>=2.6.0
redis>=5.0.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0 